[Kernel] Move the cache compiler passes into the Kernel component
